refactor(datasource): add proto feature, host file conv impls
Moves PartitionedFile / FileRange / FileGroup / FileSinkConfig conversion
impls from datafusion-proto's transient TryFromProto trait into a new
datafusion-datasource/src/proto.rs module gated by an opt-in `proto`
feature. Each impl becomes a standard `TryFrom` because the target type
is local in datafusion-datasource.

datafusion-datasource gains an optional `datafusion-proto-common` dep
behind the new `proto` feature. datafusion-proto enables it.

Round-trip test for PartitionedFile path encoding moves to its new home
in datafusion-datasource.

The `&[PartitionedFile] -> protobuf::FileGroup` impl signature changes to
`&FileGroup -> protobuf::FileGroup` (slice of foreign type can't satisfy
the orphan rule); single callsite updated.

Wire format unchanged. 161 proto tests pass (6 pre-existing
parquet-testing submodule failures unrelated).
